Revert "Feature/sol parser"
