+++ src/app/parser/impl/berlingske/BerlingskeParser.ts
@@ -20,14 +20,19 @@ import * as Cheerio from 'cheerio';
 
 import ArticleAuthor from 'base/ArticleAuthor';
 
-import AbstractLabradorParser from 'app/parser/AbstractLabradorParser';
+import AbstractIteratingParser from 'app/parser/AbstractIteratingParser';
 import IteratingParserItem from 'app/parser/IteratingParserItem';
 
 // import { getOpenGraphAuthors } from 'app/parser/util/AuthorParser';
+import { getOpenGraphModifiedTime } from 'app/parser/util/VersionParser';
 
-export default class BerlingskeParser extends AbstractLabradorParser {
+export default class BerlingskeParser extends AbstractIteratingParser {
 
 	// Implement AbstractParser
+
+	protected parseVersion() : Promise <string> {
+		return Promise.resolve(getOpenGraphModifiedTime(this.select));
+	}
 
 	protected parseByline() : Promise <ArticleAuthor[]> {
 		/*const authors = getOpenGraphAuthors(this.select);
@@ -39,6 +44,15 @@ export default class BerlingskeParser extends AbstractLabradorParser {
 
 	protected getArticleContentScope() : string {
 		return 'div#content.main-content';
+	}
+
+	protected getParsedElementNames() : string[] {
+		return [
+			'h1',
+			'h2',
+			'p',
+			'figure',
+		];
 	}
 
 	protected isMainTitle(
@@ -64,6 +78,23 @@ export default class BerlingskeParser extends AbstractLabradorParser {
 		select : Cheerio
 	) : boolean {
 		return this.isFigure(item, select);
+	}
+
+	protected isSubHeading(
+		item : IteratingParserItem,
+		select : Cheerio
+	) : boolean {
+		return item.name === 'h2'
+			&& item.text.length > 0
+			&& item.css.length === 0;
+	}
+
+	protected isParagraph(
+		item : IteratingParserItem,
+		select : Cheerio
+	) : boolean {
+		return item.name === 'p'
+			&& item.text.length > 0;
 	}
 
 	protected isFigure(
