+++ src/test/app/parser/parser-endtoend/articleSitesAndFiles.ts
@@ -5,7 +5,6 @@ import * as app from 'app/util/applib';
 export const mapSitesToParser = Object.freeze({
 	'dagbladet.no': 'Dagbladet Parser',
 	'nettavisen.no': 'Nettavisen Parser',
-	'sol.no': 'Sol Parser',
 });
 
 const htmlPath = path.join('resources', 'article', 'html');
